+++ CHANGELOG.md
@@ -7,6 +7,15 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Use of [@mischareitsma/config-parser](https://www.npmjs.com/package/@mischareitsma/config-parser)
+  to do the configuration parsing.
+
+### Removed
+
+- The configuration parsing software.
+
 ## [0.2.1] - 2024-02-13
 
 ### Changed
@@ -47,7 +56,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Initial version of logging module
 
 [Unreleased]: https://github.com/mischareitsma/logging-ts/compare/v0.2.1...HEAD
-[0.2.0]: https://github.com/mischareitsma/logging-ts/compare/v0.2.0...v0.2.1
+[0.2.1]: https://github.com/mischareitsma/logging-ts/compare/v0.2.0...v0.2.1
 [0.2.0]: https://github.com/mischareitsma/logging-ts/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/mischareitsma/logging-ts/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/mischareitsma/logging-ts/releases/tag/v0.1.0